requirements.txt
gunicorn
uvicorn
python-dotenv
requests
replicate
pillow
google-cloud-storage
streamlit
google-genai